requirements.txt
Babel==2.9.1
Pillow==8.4.0
pytz==2021.3
tinydb==4.5.2
tkcalendar==1.6.1
